[AC97] refactor, moved CMiniportStream into stream.h
